move langchain unit tests to tests folder in separate file
